+++ packages/services-http/test/content-type/json.ts
@@ -6,7 +6,6 @@ import * as querystring from 'querystring'
 import { HttpError, HttpJson } from '../../src'
 import { HttpSeed } from '../../src/mode/seed'
 use(promisedChai)
-
 interface ExpectedResponse {
   args: {
     param1: string
@@ -14,7 +13,7 @@ interface ExpectedResponse {
   }
 }
 
-@suite('Http::Json')
+@suite('Http::ContentType::Json')
 export class Test extends BasicTest {
   http = new HttpJson(new HttpSeed())
   baseUrl = `https://postman-echo.com`
@@ -48,70 +47,4 @@ export class Test extends BasicTest {
     expect(result.args).to.have.ownProperty('foo1')
     expect(result.args).to.have.ownProperty('foo2')
   }
-
-  // TODO: inject mock for following tests:
-
-  //  @test
-  //  async 'throws response when status is not 200'() {
-  //    const response = new Response('error', { status: 404 })
-  //
-  //    const fetch: (
-  //      url: string | NodeRequest,
-  //      init?: RequestInit | undefined
-  //    ) => Promise<Response> = () => {
-  //      return Promise.resolve(response)
-  //    }
-  //
-  //    const config: RequestConfig = {
-  //      init: {
-  //        headers: {
-  //          Accept: 'application/vnd.sdmx.data+json;version=1.0.0-wd',
-  //        },
-  //        timeout: 30000,
-  //      },
-  //      parseBody: 'text',
-  //      url: this.url,
-  //    }
-  //    const request = new Http.Request<Foo>(config, fetch)
-  //
-  //    request.mode = { ...request.mode, environment: 'cloud' }
-  //
-  //    try {
-  //      await request.execute()
-  //      expect.fail('request should throw error')
-  //    } catch (e) {
-  //      expect(e).to.deep.equal(response)
-  //    }
-  //  }
-  //
-  //  @test
-  //  async 'propagates the error from fetch'() {
-  //    const fetch: (
-  //      url: string | NodeRequest,
-  //      init?: RequestInit | undefined
-  //    ) => Promise<Response> = () => {
-  //      return Promise.reject(new Error('broken'))
-  //    }
-  //
-  //    const config: RequestConfig = {
-  //      init: {
-  //        headers: {
-  //          Accept: 'application/vnd.sdmx.data+json;version=1.0.0-wd',
-  //        },
-  //        timeout: 30000,
-  //      },
-  //      parseBody: 'text',
-  //      url: this.url,
-  //    }
-  //    const request = new Http.Request<Foo>(config, fetch)
-  //
-  //    request.mode = { ...request.mode, environment: 'cloud' }
-  //
-  //    try {
-  //      await request.execute()
-  //      expect.fail('request should throw error')
-  //    } catch (e) {
-  //      expect(e.message).to.equal('broken')
-  //    }
-  //  }
 }